pyproject.toml
[project]
name = "daledou"
version = "11.0.0"
description = "Q宠大乐斗脚本辅助"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "schedule>=1.2.2",
]
